# docs/openapi/v2.yaml
openapi: 3.1.0
info:
  title: LinkDominator CRM V2 API
  version: 0.3.0
  description: Production API for extension-driven LinkedIn automation using Unipile-first integrations.
servers:
  - url: https://app.linkedempire.com/api/v2
  - url: http://127.0.0.1:8000/api/v2
paths:
...
  /conversations:
    get:
      summary: List conversations
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/OrganizationIdHeader"
      responses:
        "200":
          description: Conversation list
...
  /calls:
    post:
      summary: Create call orchestration item
      security: [ { bearerAuth: [] } ]
      parameters: [ { $ref: "#/components/parameters/OrganizationIdHeader" } ]
      responses:
        "201":
          description: Created
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: opaque
  parameters:
    OrganizationIdHeader:
      name: X-Organization-Id
      in: header
      required: true
      schema:
        type: integer
...